refactor discrete functionality into discrete js files
